refactor(rls): refactoring rls edito coponents
