+++ apps/frontend/src/lib/authz/rls/RLSItem.svelte
@@ -1,13 +1,12 @@
 <script lang="ts">
-	import { rlsPolicy, type RLS } from '@undb/authz'
-	import FilterEditor from '$lib/filter/FilterEditor.svelte'
+	import type { RLS } from '@undb/authz'
 	import { Button } from 'flowbite-svelte'
 	import { t } from '$lib/i18n'
 	import { trpc } from '$lib/trpc/client'
 	import { invalidate } from '$app/navigation'
 	import { getTable, listRecordFn } from '$lib/store/table'
 	import type { IFilter } from '@undb/core'
-	import RlsSubject from './RLSSubject.svelte'
+	import RlsItemEditor from './RLSItemEditor.svelte'
 
 	export let rls: RLS
 
@@ -28,11 +27,8 @@
 </script>
 
 <li class="flex items-start gap-2">
-	<div class="flex-1 w-full p-1 rounded-md bg-gray-100 border border-gray-200 space-y-2">
-		<RlsSubject action={rls.policy.action} />
-		<FilterEditor bind:value={filter} let:add>
-			<Button on:click={add} class="w-full mt-2" size="xs" color="alternative">{$t('Create New Filter')}</Button>
-		</FilterEditor>
+	<div class="flex-1 w-full p-1 rounded-md bg-gray-100 border border-gray-200">
+		<RlsItemEditor action={rls.policy.action} bind:filter />
 	</div>
 
 	<Button class="w-20 whitespace-nowrap" color="alternative" size="xs" on:click={() => {}}>
